refactor: event 디렉토리 #148
