+++ src/main/java/com/gdg/z_meet/domain/event/controller/EventController.java
@@ -1,7 +1,7 @@
-package com.gdg.z_meet.domain.meeting.controller;
+package com.gdg.z_meet.domain.event.controller;
 
+import com.gdg.z_meet.domain.event.service.EventService;
 import com.gdg.z_meet.domain.meeting.dto.MeetingResponseDTO;
-import com.gdg.z_meet.domain.meeting.service.MeetingCommandService;
 import com.gdg.z_meet.global.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -16,13 +16,13 @@ import org.springframework.web.bind.annotation.*;
 @Validated
 public class EventController {
 
-    private final MeetingCommandService meetingCommandService;
+    private final EventService eventService;
 
     @Operation(summary = "팀 삭제 기회 추가")
     @PatchMapping("/leftDelete")
-    public Response<MeetingResponseDTO.GetMyDeleteDTO> getMyDelete(@RequestParam(name = "name") String name, @RequestParam(name = "phoneNumber") String phoneNumber) {
+    public Response<MeetingResponseDTO.GetMyDeleteDTO> patchMyDelete(@RequestParam(name = "name") String name, @RequestParam(name = "phoneNumber") String phoneNumber) {
 
-        MeetingResponseDTO.GetMyDeleteDTO response = meetingCommandService.patchMyDelete(name, phoneNumber);
+        MeetingResponseDTO.GetMyDeleteDTO response = eventService.patchMyDelete(name, phoneNumber);
 
         return Response.ok(response);
     }
